refactor(plots): move plots to separate package
Issue #81
