+++ SnowflakePlots.jl/Project.toml
@@ -1,0 +1,17 @@
+name = "SnowflakePlots"
+uuid = "fa4a7b36-de81-424e-b6b4-6a78b5e46525"
+authors = ["Alireza Yazdi <alireza@anyonsys.com>"]
+version = "0.1.0"
+
+[deps]
+Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
+Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
+PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
+Snowflake = "7bd9edc1-4fdc-40a1-a0f6-da58fb4f45eb"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
+
+[compat]
+Plots = "1.25"
+StatsBase = "0.32 - 0.33"
