+++ SnowflakePlots.jl/README.md
@@ -1,0 +1,17 @@
+![Snowflake](https://repository-images.githubusercontent.com/441460066/a4572ad1-6421-4679-aa31-4c2a45829dc6)
+
+# SnowflakePlots.jl
+
+SnowflakePlots provides plotting tools for [Snowflake](https://github.com/anyonlabs/Snowflake.jl).
+
+# Snowflake Contributors Community
+
+We welcome contributions! If you are interested in contributing to this project, a good place to start is to read our guidelines.
+
+We are dedicated to cultivating an open and inclusive community to build software for near term quantum computers. Please read our code of conduct for the rules of engagement within our community.
+
+# Alpha Disclaimer
+
+SnowflakePlots is currently in alpha. We may change or remove parts of Snowflake's API when making new releases.
+
+Copyright (c) 2023 by Snowflake Developers and Anyon Systems, Inc.
