+++ SnowflakePlots.jl/src/bloch_sphere.jl
@@ -1,27 +1,6 @@
 using Interpolations
 using PlotlyJS
 
-function plot_histogram(circuit::QuantumCircuit, shots_count::Int)
-    data = simulate_shots(circuit, shots_count)
-    datamap = proportionmap(data)
-
-    labels = String[]
-    for (key, value) in datamap
-        push!(labels, key)
-    end
-
-    Plots.bar(
-        (x -> datamap[x]).(labels),
-        xticks = (1:length(data), labels),
-        legends = false,
-        ylabel = "probabilities",
-    )
-end
-
-function viz_wigner(ρ, x, y)
-    return Plots.contour(x, y, (x,y) -> wigner(ρ, x, y), fill = true)
-end
-
 """
     BlochSphere
