+++ SnowflakePlots.jl/test/runtests.jl
@@ -1,0 +1,3 @@
+for file in readlines(joinpath(@__DIR__, "testgroups"))
+    include(file * ".jl")
+end
